Require pygpu 0.6.5 or greater
Switch to requiring `pygpu` instead of `libgpuarray` as `libgpuarray`
will become the C only library and `pygpu` the Python frontend. Also
require `pygpu` 0.6.5 or greater because this is the first version that
is packaged in `conda-forge`. Older versions come only from `defaults`
currently and the build `pygpu` differently. So this is the first
version we can be sure of working as intended.